03_openai_agent_sdk/01_agents/pyproject.toml
[project]
name = "01-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.0.16",
]
